Не правильно считались проценты роста. Поправил

+++ src/main/java/com/antonromanov/arnote/service/investment/returns/ReturnsService.java
@@ -17,7 +17,7 @@ public interface ReturnsService {
      * @param user
      * @return
      */
-    Optional<Long> getSharesDelta(ArNoteUser user);
+    Optional<Double> getSharesDelta(ArNoteUser user);
 
     /**
      * Получить общую доходность по дивидендам.
